split up code in CellularAutomataMusic.java and placed the code in  a new controller
